requirements.txt
django
djangorestframework
django-cors-headers
pyyaml
djangorestframework-simplejwt
